requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
load-dotenv==0.1.0
python-dotenv==1.0.1
python-telegram-bot==21.8
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
yt-dlp==2024.12.3

aiohttp~=3.11.9
validators~=0.34.0
instaloader~=4.14